Chore:wrench:: #36 Add Get Models
