--- GiwazipClient/Models/Model.swift
@@ -1,8 +1,0 @@
-//
-//  Model.swift
-//  GiwazipClient
-//
-//  Created by 김민택 on 2023/01/13.
-//
-
-
